src/app.py
import os
from sqlalchemy import create_engine
import pandas as pd
from dotenv import load_dotenv

# load the .env file variables
load_dotenv()

# 1) Connect to the database here using the SQLAlchemy's create_engine function
connection_string = f"postgresql://{os.getenv('DB_USER')}:{os.getenv('DB_PASSWORD')}@{os.getenv('DB_HOST')}/{os.getenv('DB_NAME')}"
DB = create_engine(connection_string).execution_options(autocommit=True)
DB.connect()

dataframe = pd.read_sql("Select * from publishers", DB)
print(dataframe)
